+++ data-raw/ebird-taxonomy.r
@@ -15,7 +15,7 @@ ebird_taxonomy <- read_xlsx("data-raw/eBird_Taxonomy_v2017_18Aug2017.xlsx") %>%
   as.data.frame(stringsAsFactors = FALSE)
 
 write_csv(ebird_taxonomy, "data-raw/ebird-taxonomy.csv", na = "")
-devtools::use_data(ebird_taxonomy, overwrite = TRUE, compress = "xz")
+usethis::use_data(ebird_taxonomy, overwrite = TRUE, compress = "xz")
 
 
 filter(ebird_taxonomy, scientific_name == "Gyps rueppelli")