+++ NEWS.md
@@ -4,6 +4,7 @@
 - New object, `auk_sampling`, and associated methods for working with the sampling data only
 - New function, `auk_select()`, for selecting a subset of columns
 - `auk_date()` now allows filtering date ranges across years using wildcards, e.g. `date = c("*-05-01", "*-06-30")` for observations from May and June of any year
+- New function, `auk_state()` for filtering by state
 - Now using AWK arrays to speed up country and species filtering; ~20% speed up when filtering on many species/countries
 - Allow selection of a subset of columns when filtering
 - Remove free text columns in `auk_clean()` to decrease file size